+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Notable changes to this project.
 ## [] -
 - fix downloading dataset for tournaments > 1
 - add `validationCorrelation` and `liveCorrelation` to all relevant places
+- remove `validationAuroc` and `validationLogloss` from `submission_status`
 
 ## [1.6.0] - 2019-07-10
 - default to tournament 8 `katzuagi`